SpiSlave, _shift: add interruptible option
This allows the _shift operation to be stopped at any time through a
deassert of CS. If this option is set, a status indicating if the
transaction was stopped prematurely is additionally returned.
This also checks for premature stop of the transaction at each clock
edge, resulting in an error message precisely when the abort happens
in the original case.